reports: open reports via webadmin - host or vm gives exception (#855865)
https://bugzilla.redhat.com/855865

Exception tends to occur in the reports integration.
The details of this exception is:

org.ovirt.authentication.EngineUserDetails
cannot be cast to
com.jaspersoft.jasperserver.api.metadata.user.domain.User

This shouldn't happen, as jasper is supposed to convert the
EngineUserDetails from the token to its own user details, and leave the
original details in originalAuthentication field.

However, in this patch, when re-authentication is needed, we leave the
current authentication object, setting only our user details. That way,
jasper doesn't need to recreate their own user details every time.

Change-Id: Ie9c2ce53018e2b9d08edb5d0162c584c81ac03d5
Signed-off-by: Oved Ourfali <oourfali@redhat.com>