packaging: setup: Recreate dwh data source on upgrade
Do not export/import dwh data source on upgrade. Instead, always create
it as on a clean setup.

This fixes the flow in the linked bug, where a user tries to restore a
backup that includes dwh and reports to a dwh database with different
credentials.

In principle it might break manual changes done to the datasource after
running setup.

Change-Id: I1fff0f30991dd839921a9e2af7f6b3e5290370c7
Bug-Url: https://bugzilla.redhat.com/1229398
Signed-off-by: Yedidyah Bar David <didi@redhat.com>